+++ e2e/tests/plugins.e2e.js
@@ -231,6 +231,7 @@ describe('plugins', () => {
       cy.shrinkPlugin();
       cy.waitForVideoToLoad();
       cy.focusEditor()
+        .type('{uparrow}') //try to fix bug where sometimes it doesn't type
         .type('{uparrow}')
         .type('Will this fix the flakiness?');