+++ packages/plugin-gallery/web/src/components/gallery-controls/layouts-selector.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 
-import { getGalleryLayouts } from '../../helpers';
+import { getGalleryLayouts } from '../../layout-helper';
 import { mergeStyles } from 'wix-rich-content-common';
 import { SelectionList } from 'wix-rich-content-editor-common';
 import styles from '../../../statics/styles/layout-selector.scss';