+++ packages/plugin-mentions/web/src/editor.js
@@ -1,6 +1,6 @@
 import { createExternalMentionsPlugin } from './createMentionsPlugin';
 import { MENTION_TYPE } from './types';
-import { DEFAULTS } from './defaultSettings';
+import { DEFAULTS, THEME as theme } from './defaultSettings';
 
 export const pluginMentions = (config = {}) => {
   return {
@@ -8,5 +8,6 @@ export const pluginMentions = (config = {}) => {
     type: MENTION_TYPE,
     createPlugin: createExternalMentionsPlugin,
     ModalsMap: {},
+    theme,
   };
 };