+++ packages/wrapper/web/src/localeStrategy/localeStrategyProvider.js
@@ -4,9 +4,9 @@ export default ({ locale = 'en' } = { locale: 'en' }) => async (innerProps = {})
     return innerProps;
   }
   try {
-    const locale = `wix-rich-content-common/dist/statics/locale/messages_${mergedLocale}.json`;
-
-    const localeResource = await import(locale).then(res => res.default);
+    const localeResource = await import(
+      `wix-rich-content-common/dist/statics/locale/messages_${mergedLocale}.json`
+    ).then(res => res.default);
     return { locale: mergedLocale, localeResource, ...innerProps };
   } catch (err) {
     throw new Error(`error while loading locale ${locale}`, err);