+++ scripts/templates/atomic-plugin/web/src/editor.js
@@ -1,0 +1,19 @@
+/*
+  This module exports the wrapper neccesities for RichContentEditor.
+  If your plugin uses a modal, then make sure to uncomment 'ModalsMap'.
+*/
+
+import { createYourPluginNamePlugin } from './createYourPluginNamePlugin';
+import { YOUR_PLUGIN_NAME_TYPE } from './types';
+// import { ModalsMap } from './modals';
+import { DEFAULTS, THEME as theme } from './defaults';
+
+export const pluginYourPluginName = (config = {}) => {
+  return {
+    config: { ...DEFAULTS.config, ...config },
+    type: YOUR_PLUGIN_NAME_TYPE,
+    createPlugin: createYourPluginNamePlugin,
+    // ModalsMap,
+    theme,
+  };
+};