+++ scripts/templates/atomic-plugin/web/src/index.js
@@ -1,3 +1,4 @@
 export { createYourPluginNamePlugin } from './createYourPluginNamePlugin';
-export { default as yourPluginNameViewer } from './yourPluginName-viewer';
+export { default as yourPluginNameViewer } from './yourDpluginDname-viewer';
 export { YOUR_PLUGIN_NAME_TYPE } from './types';
+export { pluginYourPluginName } from './editor';